+++ NEWS.md
@@ -1,5 +1,7 @@
 # cpp11 (development version)
 
+* Improve consistency of inserting NA values in r_string objects (#45)
+
 * Added a `NEWS.md` file to track changes to the package.
 
 # cpp11 0.1.0